Refs #13487: apply suggestions
